Huge Commit. Read description for further Information
Added another Table to the current view:  StockChangeLog.

Changes to the Stock will be documented and saved in this Table with Timestamps

This new Table is a addition to the used article Database. Migration to an own DB is possible but very very hard and time consuming. MB a temporary solution